add option to CMake to selectively utilize the legacy (non-RTP) network protocol (embedded FNE does not support this and will be disabled for USE_LEGACY_NETWORK; cleanup some error messages when trying to initialize and read the configuration file; fix some minor issues in parsing the new talkgroup rules YAML file; complete conversion of BaseNetwork from raw socket to using the new FrameQueue class; refactor the multi-buffer UDPSocket write function to properly send multiple messages in batch down to the kernel driver (on Linux using sendmmsg); add flag to UDPSocket to indicate whether the socket is "open" or not; correct some RTP formatting issues with the extension header (the payload length is in 32-bit units);

+++ src/lookups/TalkgroupRulesLookup.h
@@ -54,12 +54,11 @@ namespace lookups
         }
         /// <summary>Initializes a new insatnce of the TalkgroupRuleGroupVoiceSource class.</summary>
         /// <param name="node"></param>
-        /// <param name="tgSlot"></param>
         TalkgroupRuleGroupVoiceSource(yaml::Node& node) :
             TalkgroupRuleGroupVoiceSource()
         {
             m_tgId = node["tgid"].as<uint32_t>(0U);
-            m_tgSlot = node["slot"].as<uint8_t>(0U);
+            m_tgSlot = (uint8_t)node["slot"].as<uint32_t>(0U);
         }
 
         /// <summary>Equals operator. Copies this TalkgroupRuleGroupVoiceSource to another TalkgroupRuleGroupVoiceSource.</summary>
@@ -99,8 +98,6 @@ namespace lookups
         }
         /// <summary>Initializes a new insatnce of the TalkgroupRuleConfig class.</summary>
         /// <param name="node"></param>
-        /// <param name="affilated"></param>
-        /// <param name="routable"></param>
         TalkgroupRuleConfig(yaml::Node& node) :
             TalkgroupRuleConfig()
         {
